Fix: Use a handrolled fgets to read from target process.
This fixes a rather awkward problem where you try to read a string from the
target process and it happens to be at the end of one of their address space
regions. In this scenario you would hit EOF and fgets would return NULL, which
we interpret as failure. It seems hard to determine whether fgets has hit this
or an immediate EOF, indicating an invalid pointer on the part of the target.