Interpose on std{in|out|err} as well.
We need to track these in order to replicate them when restoring from the
cache. Note that this commit does nothing with the results of this tracking
yet. We've also had to switch to fork since we're dup2-ing before the execvp.
This is not a big deal because we can avoid it if necessary by dup2-ing before
the fork/vfork, and then restoring our (the parent's) std* afterwards.