Chmod the copy source either side of a cp operation.
This works around a situation where the file we're copying from is not
currently readable. Note that this also drops any attempt to retain permissions
on the destination. We need a new strategy here as I just realised the hashes
of files containing identical content, but with different permissions, collide.