Handle multiprocess operations.
This changeset represents a fairly large and chaotic refactor. Unfortunately
the assumption that the operation we're tracing was a single process was fairly
solidly baked into a lot of cases. This change has not been done particularly
cleanly and has some significant problems that should be fixed ASAP including,
but not limited to:
 - Clients need to be too careful when dealing with tracees. E.g. when giving
   up, you need to manually acknowledge the last syscall you received before
   detaching. The essential problem here is that trace.c does not track when
   processes are blocked.
 - Almost every ptrace call site was a problematic location in this change. The
   ptrace API is really showing its warts and we should abstract it to avoid
   some of the gotchas.
 - The 'syscall' type should track which process (not just which PID) it is
   associated with. By doing this, we could have some understanding of which
   processes are currently halted.