xxxcache: Beginnings of Python prototyping tool.
This is something that has been on the todo list for awhile. It is intended to
be a simpler version of xcache for a constrained program where you know
something about its inputs and outputs. The idea is that you can use this
during prototyping, debugging, or when you simply just don't need the full
power of xcache.