Replace internal list implementation with GLib's SList.
It's not worth maintaining our own implementation of linked-lists, despite
GLib's unpleasant semantics. This changeset leaves list.{h|c} in place as a
wrapper in case we want to switch list implementation's later. Note that this
changeset also exposes the list_t type in the header for caller allocation if
preferred.

The other generic structures, dict and set, should be migrated to GLib as well.

Getting GLib to work with CMake was an upsetting little experience.