Update process state when unblocking.
In theory this should not be necessary as we only ever manually unblock a
process like this when we're about to detach and let it run free. Still, it
helps to be consistent here. The right solution long term may actually be to
introduce a new state for 'DETACHED' to indicate when we're no longer
monitoring the state of a process.