xcache: Avoid tracking O_CREAT|O_TRUNC and O_CREAT|O_EXCL files as inputs.
This copes with some cases in GCC where it opens files O_RDWR and includes
these flags. It goes on to use its read-write access to manage the output
in-place. So, despite opening the files O_RDWR, they are used exclusively as
inputs.

I was hoping this would resolve the access pattern to a.out, but unfortunately
it is stated, lstated and unlinked before even opening. It's not immediately
clear to me why it can't just blindly open it O_RDWR|O_CREAT|O_TRUNC, but I
suspect this is related to security or checking free space prior.