Reimplement `asprintf` internally.
This puzzles me, but `asprintf` seems to unconditionally leak memory. Some
googling suggests this is a known issue and nobody cares. Given that the number
of times we invoke `asprintf` is dependent on the target we're tracing, we leak
an unconstrained amount of memory. This makes it a little unsuitable for
xcache.