Trace the children of children as well.
It is not clear to me that this is necessary (do PTRACE_O_TRACEFORK and friends
act recursively or only on the immediate child?). However, it seems prudent to
apply it for now in pursuit of troubleshooting Bitbucket issue #6.
Unfortunately this does not resolve the GHC issues, but has helped shed some
light on the situation.