Fix: tools/strace: Always set PTRACE_O_TRACEEXEC.
Buried somewhere in the ptrace docs is the note that, if you don't set this
option, the tracee will receive a spurious SIGTRAP when it calls execve. This
signal seems intentional, but it is not clear what the purpose of it is, as
tracers are instructed to suppress it. Using PTRACE_O_TRACEEXEC allows us to
avoid incorrectly forwarding SIGTRAP to the tracee and killing it, at the
expense of copping tracing events we don't care about.

This still doesn't solve the GHC woes, though it does let us trace it without
followforks.