tools/strace: Finally fix GHC hang.
I'm not sure I can overstate how unsatisfying it was to finally get to the
bottom of this one. While comparing to the `strace` code base as part of my
detective work, I noticed that it was using wait4(), rather than wait(). This
stuck out because wait4() is deprecated, but it was also passing this
mysterious flag `__WALL` that it defined in-tree with no comment. Way down the
bottom of the wait() manpage there is a discussion of Linux-specific flags and,
in particular, how they affect thread groups. Anyway, the upshot of all of this
is that you need to pass `__WALL` (also defined in <sys/wait.h>) to actually
wait on _all_ children. Making this tweak magically resolves our GHC issues.
Slight grrr.