xcache: Fix: Wait on non-clone processes as well.
This fixes the blocker on Bitbucket #6, but now we hit another issue. We
attempt to bailout from GHC due to an unhandled mkdir(), but then we deadlock
while detaching. Comparing our detach logic to the extremely complex `strace`
detach logic, I suspect we are not doing a thorough enough job here.