xcache: Fix failure to detach from GHC.
Again, this was related to thread groups.

The usage of tkill() here looks quite unsafe; in particular, the tkill()
manpage advises that you may end up signalling the wrong thread if the thread
ID has been reused. I think this should be acceptable because we ~know the
thread ID is still live. This is also what `strace` does and I'm not aware of
anyone reporting this problem there.

Closes Bitbucket #6. Ding dong the witch is dead.