Fix: Resolve paths correctly during *at syscalls.
The previous code incorrectly called `normpath` when the active path might
already be absolute. That is, it is legal to call something like:

    openat(4, "/my/absolute/path", O_RDONLY)

Not only was this incorrect, but it caused `normpath` to perform OOB writes as
its first argument was typically much smaller than `PATH_LEN`. Interestingly
this was identified not by Valgrind but Glibc's malloc debugger.