Fix: Test for column type of environment values.
It turns out SQLite actually tests the data type of the value in the current
row rather than any "column type" when you call `sqlite3_column_type`. This
means that the column reports as `SQLITE_NULL` if the row we're looking at
happens to contain `NULL` in this column. This can quite easily happen with a
nullable field like `value` in `env`.