dup /dev/null over stdin instead of a pipe to the parent
On consideration, it does not make sense to also intercept stdin. We do not know
the interleaving of bytes on stdin/stdout/stderr (at least we were not planning
to track this), so replaying will behave erratically in the presence of
interactive input. The simplest way to avoid this is just disallow tracing
anything that uses stdin. This also avoids some upcoming complication due to one
of the inputs being a FIFO.