use 'waitpid' instead of pidfd polling for monitoring
It seems pidfds are not usable for intercepting trace events. The setup goes
fine, but the pidfd does not become readable when trapping events occur. This
makes it basically unusable for our purposes. This is unfortunate, as it means
we need to once again resort to a multithreaded tracer.

This change also moves away from `waitid` to `waitpid`. There are mumblings on
the internet about `waitid` being unusable for tracing because it mangles the
status return that makes retrieving the extra ptrace bits impossible.