fix bad 'ptrace' setoptions call
This call was passing one too many parameters to `ptrace`, something the
compiler cannot easily detect due to `ptrace`’s varargs API. The effect of this
typo was that this call was a no-op. It is incredible that this was not detected
earlier as this has a significant effect on the semantics of the program. I
guess a “barely there” tracer is surprisingly effective on trivial programs.