manager tracee setup entirely via 'wait'
The tracer/tracee interactions during setup were written based on the
exec-with-close-on-exec-fd. However an interesting wrinkle to our usage is that
the thing we are planning to do immediately after execing the tracee is to
`wait` on it. We can take advantage of this by purely managing interactions with
the tracee through `wait`. This significantly simplifies the situation; perhaps
not obvious from this commit alone.