trace any unrecognised syscall
A simplification that I had not previously realised: ignoring a known “safe” set
of syscalls and then SIGTRAPing everything else is both clearer and more
efficient. This means learning any new syscall involves either:

  1. Marking it as safe by ignoring it in the seccomp filter; or
  2. Teaching the tracer how to interpret it.

Note that this required moving the SIGSTOP synchronisation _prior_ to installing
the seccomp filter. This is fine (it is to synchronise the `PTRACE_ME`, not the
seccomp filter install), but I am still not sure why it is necessary. Leaving it
after the seccomp filter installation resulted in (1) the tracee seeing a -1
return to their `raise` call and (2) the tracer seeing a SIGSEGV from the
tracee. Neither of these are accurate, so I am not sure what is happening here.
I suspect it is currently impossible for the tracee to `raise` without confusing
the tracer. This will need to be root caused in future.

Note that a side effect of this is that the tracer now sees `proc_exec`’s
`chdir`. I had initially thought of suppressing this, but I realise now that the
tracer actually _needs_ to see this call in order to have an accurate
understanding of the tracee’s cwd.