fix Linux kernel version checks
It seems `LINUX_VERSION_MINOR` does not exist. This was apparently never noticed
because this code has not been compiled on Linux 3.x. In future we should audit
both the compile-time and runtime checks in the code and CMake because they seem
overlapping in a way that is not entirely consistent.