improve how Linux kernel versions are handled
The code attempted to cope with three different Linux kernel version ranges:
  1. Linux < 3.5
  2. Linux [3.5, 4.8)
  3. Linux ≥ 4.8
However the build system rejected any kernel from (1) or (2). I cannot remember
if this was based around some idea of supporting cross-compilation or if this
was just an accidental inconsistency.

This change does something more rigorous. Compilation results in a binary that
supports anything less than or equal to your current version. At runtime,
anything too old or new for your current version is disabled. This seems to get
us the best possible outcome.