remove unnecessary duplicate 'procs' array reservation
While working on simplifying the `procs` array, it was discovered that this
reservation is redundant. It reserves space for a new `procs` entry, but the
upcoming `inferior_start` call in `monitor` does that anyway. The parent
(`xc_record`) already needs to anticipate failure of `monitor`’s
`inferior_start` call, so this is a reasonable simplification.