fix inability to discriminate file-not-found when running the tracee
When trying to trace a non-existent program, the result would be a confusing
situation where it appeared the tracee had tried to open a file that did not
exist. In a sense, this was accurate (the file that did not exist was _itself_).
But this was confusing to users.

This change effectively introduces another return value to `xc_record`. I do not
particularly like the form of this solution – there are too many moving parts
and too many implicit non-local assumptions – but I cannot immediately think of
a better approach.

In future, we should:
  1. Unify the 3 return values of `xc_record` somehow; and
  2. Remove the footgun that the caller has to initialise `exec_status` and
     `exit_status`.