split 'xc_record' return indicators
`xc_record` was conflating a number of different things into three return values:
  1. Whether `xc_record` itself ran into incorrect caller parameters (`EINVAL`),
     resource exhaustion, etc.
  2. Whether the child failed to exec to begin with.
  3. Whether something like resource allocation failed in the monitor subthread.
  4. Whether the child did something unsupported.
  5. Whether the child exited with failure.

This was not inherently a problem for `xc_record` itself. But callers could not
easily distinguish these situations. In particular, the child’s exit status was
routed to whether the trace was saved rather than `xcache`’s eventual exit
status. This had two main negative effects:
  1. Recording a trace for a child that exited with non-zero exit status was not
     possible. This is a use case we intend to support.
  2. `xcache` would exit with a differing exit status than a traced child that
     exited with non-zero exit status. This breaks the illusion that `xcache`
     (even when tracing an unsupported child) is transparent.

This change splits the return values into four that are more cleanly separated.
I am still not very happy with this. It is too easy to confuse which statuses
are set when, and whether the responses to errors within the monitor subthread
are appropriate. Hopefully this can be improved and simplified in future.

This change un-xfails some test cases where the test logic now cannot _see_ that
recording is failing.