invert proc→thread relationship
This change flips how processes and threads are dealt with. Instead of a process
owning a collection of threads, multiple threads now own a process. Conceptually
this seems like a step backwards. However, the advantage of this change is that
we can now think unambiguously in terms of threads. Any time we need a thread’s
process, it is available through the `proc` member. But otherwise we can stick
to dealing with the most authoritative identifier.

Note that this also fixes some other warts. E.g. a process exit from `waitpid`
is now correctly recorded as a thread exit. Similarly, we kill all threads in a
process instead of killing all processes.

The exit status of a tracee is considered to be the exit status of its first
thread. I believe this is consistent, except perhaps for programs that
daemonize. This might require more thought.