handle clone sysenter
When a thread calls `clone`/`clone3`/`fork`/`vfork`, when using ptrace the
tracer gets the sequence of events:
  1. `clone` etc sysenter
  2. clone
  3. `clone` etc sysexit

This change handles (1). We now should be able to successfully trace `clone` and
`clone3`. `fork` and `vfork` are only handled when the libc call results in a
`clone` or `clone3` syscall. If `fork` and `vfork` actually map to the syscalls
of their name, more work is needed.

This required adjusting the xfail for the forking test because, while we are now
able to trace `forker`, we seem to not record the child’s actions correctly.