handle clone events
When a thread calls `clone`/`clone3`/`fork`/`vfork`, when using ptrace the
tracer gets the sequence of events:
  1. `clone` etc sysenter
  2. clone
  3. `clone` etc sysexit

This change handles (2). We are still not handling (1) or (3), so attempts to
actually do this fail recording. But this should set us up to implement this in
future.

Note that this actually explains the confusion I previously had in the removed
comment. We _do_ need to observe the clone event because it seems to be the only
way to associate the newly spawned child with its creator/parent.