fix: never detach threads
Detaching threads when tracing was recognised to have failed was assumed to be
the most efficient approach. While this may be true, it creates an awkward
situation where xcache is no longer precisely sure of the state tracee threads
are in. Some may be traced, some may not. The ones that are not may do
(unobserved) actions that create or destroy threads. The end result is that the
tracee’s state can become completely out of sync with xcache’s idea of its state
leading xcache to take incorrect code paths.

To simplify handling of this, lets just unconditionally trace threads to the
tracee’s eventual completion. This may be less efficient, but at least we can
better manage xcache’s knowledge of the tracee.

This is almost certainly not the cleanest form of this new design and could be
improved upon in future.