fix tracking of thread’s state when syscall handling fails
By continuing too early here we would fail to notice that a child thread was in
a syscall after `sysenter`. The effect of this would be that the sysexit from
this call resulted in another call into `sysenter` instead of a call into
`sysexit`. Xcache would then get thoroughly confused about the tracee’s state.