include mode within saved 'open' calls
Previously an `open` with `O_CREAT` was modelled as an `open` followed by a
`chmod`. This works in some limited cases but breaks down when considering (1)
umasks and (2) atomicity. It is not totally clear yet whether we need to take on
both of these as requirements but lets conservatively plan that we do for now.