update record of file creation mode during finalisation
Given the outputs of a trace are applied in a single pass during replay, we may
as well just record the final mode of created files instead of tracking what
specific mode they were originally created with. This avoids the need to track
umasks and will also allow this finalisation to “dominate” any intervening
chmods.