fix uncacheable test
It turns out the Glibc `getrandom` wrapper caches some data to reduce the number
of `getrandom` syscalls. A side effect of this is that our changes to probe
`mktemp` in the spy that transitively call `getrandom` caused the `getrandom` in
this test to no longer perform an actual syscall.

This change moves to a more robust implementation where we directly force the
syscall.