use an invalid file descriptor instead of pipe for spy→tracer messaging
A pipe was being constructed and then passed to the tracee for use by
libxcache-spy to message back to the tracer. This worked but had a few warts:
  • We were not being particularly careful about what `ioctl` values were used
    as message IDs. It is possible (though none are known) that some of these
    collided with actual valid `ioctl`s on pipes. This would mean the spy’s
    messages had unintended (though probably benign) side effects.
  • The pipe may have been marked close-on-exec. I have not precisely traced
    whether this was the case. But if so, it would mean that execed children of
    the tracee would potentially reuse the pipe’s file descriptor for some other
    valid file.
  • Despite this being a valid pipe, neither the tracer nor the tracee ever
    wrote anything to it or read anything from it. It was only used to perform
    `ioctl`s on.
  • The spy’s actions do not depend on the success of the `ioctl`.
It is this last point which revealed this change is possible: we can simply not
open any file or pipe at all and just let the spy operate on a known invalid
file descriptor.