fix missing dependency on libxcache-spy
This might explain some of the unexpected GCC stalls I witnessed in the past. I
was probably recompiling with `cmake --build build --target check`, so
libxcache-spy would not necessarily have been rebuilt. Running an xcache and
libxcache-spy that were out of sync would probably result in all kinds of
unexpected behaviour. In future, we should introduce more of a safety net to
catch this mismatch.