handle 'putenv'
It is unclear how to handle the string passed to `putenv` being modified after
the call. But I have never seen code that does this in the wild. So lets assume
this is a rare case we can avoid not to handle.

It is also unclear how to handle direct modification of `environ`. I have not
come across code that does this in the wild either. So lets assume this is rare
for now too.

That means the only remaining environment aspect to handle is `clearenv`.