handle 'clearenv'
This was surprisingly subtle to handle. My initial thought had been to walk the
entire environment and add every variable to the `env` set. But this results in
a situation where future `getenv` of a variable that was _not_ set is
incorrectly determined to still have an external origin. Thankfully this
implementation of using a “universe” seems to both handle all cases and be more
performant than my first idea.