libxcache-spy: outline 'sysconf' interposing
This is slightly more scalable, in that we now only interpose on `sysconf` if it
is ever used. Otherwise we do not pay this overhead.

Note that we need to handle some complications around multi-threading now. In
future, we may need to get more rigorous about accounting for recursion when
hitting `CALL_OFF` points.