make ASan usable
It turned out some of the Address Sanitizer problems we were running into before
were due to a much more subtle reason than I realised: the spy was built with
ASan while the tracee typically was not. This resulted in pre-loading an
ASan-dependent library into a non-ASan-instrumented process, hence the warnings
and failures.

With this change, we can now run xcache with ASan enabled and diagnose some
latent problems.