disable sanitizers in the test binaries
On reflection, trying to support record and replay of ASan-instrumented programs
seems like a fruitless quest, hence reverting the prior work towards this. ASan
inherently performs some unreplayable actions (e.g. calling `clock_gettime`) and
it is not obvious why you would _want_ to xcache an ASan-instrumented program.

Perhaps, as a nicety in future, we could detect when a tracee is
ASan-instrumented and not inject the spy. This would at least avoid confusing
“libasan was not loaded first” errors and allow execution of the tracee that
would then ultimately fail recording. This seems like the least surprising
behaviour we could aim for.

This change repairs almost all the test suite under ASan.