get FD paths via /proc instead of tracking
A previous iteration of xcache read the /proc file system whenever it needed to
answer “what path does this file descriptor point to?” The implementation prior
to this commit instead mirrored the file descriptor tables of the tracee’s
threads in its tracking data structures. The motivation for switching approaches
here was that it was assumed mirroring and then constructing paths in xcache’s
address space would be cheaper than taking a `readlink` every time we need to
resolve a path. However this overlooked a number of things:
  1. We are always in the midst of a ptrace stop when dealing with these
     questions. So the cost of an extra `readlink` is unlikely to be
     significant.
  2. Mirroring the file descriptor tables needs to be done _exactly_. In
     particular, if the kernel we are running on has a way of creating file
     descriptors that we are not aware of, the file descriptor mirror can become
     out of sync with reality. This results in xcache getting confused when
     trying to construct paths based on a file descriptor it believes to not be
     open. At this point, tracing has usually failed due to seeing the unknown
     call that created the rogue file descriptor. But mirroring is still
     happening as tracing proceeds to its eventual failure.
  3. Mirroring, though perhaps saving a `readlink`, actually incurs extra
     overhead in the form of processing syscalls we could otherwise ignore. See
     the changes related to `close` in this commit, for example.

This change switches back to resolution via /proc. This is not only vastly
simpler, but more guaranteed-correct. We no longer need to worry about whether
we are precisely mirroring kernel behaviour.

This introduces a dependency on /proc being mounted, which is not guaranteed.
But an unmounted /proc is a relatively rare configuration and we were arguably
already depending on it being mounted due to how we find the spy.