store a synthetic path for 'getenv' inputs
This looks a bit weird; we are now storing duplicated sources of truth when
tracking `getenv` actions. However this removes an important foot gun.
Previously when dealing with input paths, unless you were aware of the wrinkle
that `getenv` inputs store not a path but the environment variable name in
`path`, you could easily write invalid logic with no compiler warnings. The
“path” for such inputs is now a valid (though implementation-defined) path. This
will allow some simplification when dealing with dependencies ahead.