replace and unify de-duping optimisation logic
As discussed in baddbbc8e85a9887b3ff6c5fd84844d1e095cb07, it has become clear
that the logic for detecting duplicate or redundant operations needs to become
sophisticated enough that it will begin to look like a compiler’s optimiser.
Rather than keep this as a set of one-off rules, this change centralises the
logic into what will hopefully be an extensible design. We can now consider,
more holistically, what constitutes valid reorderings.

The rules implemented in this change are a superset of the previous logic.
Though they are still unlikely to be enough (yet) to successfully trace more
complex programs.

The workhorse example that we are making our way towards is something that
creates temporary files during its execution:

  open("/tmp/foo", O_CREAT|O_EXCL…, …);
  open("/tmp/foo", O_RDONLY);
  unlink("/tmp/foo");

This will probably require a new category of dependency that captures the
relationship between `open(…, O_CREAT|O_EXCL…, …)` and `unlink(…)`. Namely that
the latter “undoes” the former, and thus both can be dropped.